fix: only delete unused thumbnails

+++ suite/slides/doctype/presentation/presentation.py
@@ -21,15 +21,17 @@ class Presentation(Document):
 		for slide in self.slides:
 			if slide.thumbnail and slide.thumbnail.startswith("data:image"):
 				old_thumbnail = old_slides[slide.idx - 1].thumbnail
-				delete_old_thumbnail(old_thumbnail)
+				delete_old_thumbnail(slide.name, old_thumbnail)
 				slide.thumbnail = save_base64_thumbnail(slide.thumbnail, self.name)
 
 	def validate(self):
 		self.update_thumbnails()
 
 
-def delete_old_thumbnail(old_thumbnail: str):
+def delete_old_thumbnail(slide_id: Document, old_thumbnail: str | None = None):
 	if old_thumbnail and old_thumbnail.startswith("/private/files/"):
+		if frappe.db.exists("Slide", {"thumbnail": old_thumbnail, "name": ["!=", slide_id]}):
+			return
 		try:
 			file_doc = frappe.db.get_value("File", {"file_url": old_thumbnail})
 			frappe.delete_doc("File", file_doc)
@@ -127,7 +129,8 @@ def insert_slide(name, index):
 @frappe.whitelist()
 def delete_slide(name, index):
 	presentation = frappe.get_doc("Presentation", name)
-	delete_old_thumbnail(presentation.slides[index].thumbnail)
+	slide = presentation.slides[index]
+	delete_old_thumbnail(slide.name, slide.thumbnail)
 	presentation.slides = presentation.slides[:index] + presentation.slides[index + 1 :]
 	for i in range(index, len(presentation.slides)):
 		presentation.slides[i].idx -= 1
